backend/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
pymongo==4.6.2
python-engineio==4.9.0
python-socketio==5.11.1
eventlet==0.35.2 